+++ tests/integration.rs
@@ -1,0 +1,10 @@
+//! Integration tests for blazesym.
+
+#![allow(
+    clippy::fn_to_numeric_cast,
+    clippy::let_and_return,
+    clippy::let_unit_value
+)]
+#![cfg_attr(not(linux), allow(dead_code, unused_imports))]
+
+mod suite;
