+++ tests/suite/mod.rs
@@ -1,0 +1,3 @@
+mod inspect;
+mod normalize;
+mod symbolize;
